pyproject.toml
[project]
name = "ghost-assistente"
version = "0.1.0"
description = "Assistente por voz"
authors = [
    {name = "retr0",email = "kg3139120@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.4.4,<3.0.0)",
    "speechrecognition (>=3.16.0,<4.0.0)",
    "pyaudio (>=0.2.14,<0.3.0)",
    "piper-tts (>=1.4.2,<2.0.0)",
    "psutil (>=7.2.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
